website/website/requirement.txt
python==3.6.9
Django==2.1
django-cors-headers==3.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
httplib2==0.9.2
PyJWT==1.7.1

